// course/server/src/constants/types.ts
export enum TableRoutes {
  accounts = '/accounts',
  clients = '/clients',
  comments = '/comments',
  couriers = '/couriers',
  managers = '/managers',
  orders = '/orders',
  pizzas = '/pizzas',
  reports = '/reports',
  statuses = '/statuses',
}

export enum TableNames {
  accounts = 'Accounts',
  clients = 'Clients',
  comments = 'Comments',
  couriers = 'Couriers',
  managers = 'Managers',
  orders = 'Orders',
  pizzas = 'Pizzas',
  reports = 'Reports',
  statuses = 'Statuses',
}